Add JSON tests and external types.
Summary: The ExtTypes module has been added is intended to contain types that
could trivially support isomorphic to/from JSON functions. Those JSON
generation functions have been added in the ExtJSON module. The JSON
module has been mostly untouched. The plan is to move all of the ExtJSON
code into the JSON module and consolidate the ExtTypes and Types
modules. The Treb.Test.Gen module contains all of the QuickCheck
Arbitrary instances of the types in the ExtTypes module. Some of these
need improvement to have properties that we expect, e.g. fields in a DataBlock
should have unique names. JSON schemas have been defined in
Treb.Test.Schema using the json-schema package, covering the types in
the ExtTypes module. It lacks many validation features of JSON schema, but
at least covers many of the key structural properties. Treb.Test.JSON
allows for a test of a specific (To|From)JSON instance to be specified
by giving the type. Many of these (To|From)JSON instances have been
tested and succeed, but they need to be revisited soon as some things
have changed.

Test Plan: ExtJSON, ExtTypes have been tested and those which haven't
been tested will be soon. The framework is there so that one must only
tell the testing tool which types you want to investigate and the test
is defined.

Reviewers: travis